更新服务器间连接可能会出现的一些警告代码。

+++ trunk/ purenessscopeserver --username freeeyes1226@gmail.com/PurenessScopeServer/Reactor/ConnectClient.cpp
@@ -241,7 +241,7 @@ bool CConnectClient::SendData(ACE_Message_Block* pmblk)
 		}
 		else
 		{
-			nDebugSize = pmblk->length();
+			nDebugSize = (int)pmblk->length();
 		}
 
 		char* pData = pmblk->rd_ptr();
